requirements.txt
pandas>=2.2
numpy>=1.26
requests>=2.32
python-dotenv>=1.0
geopandas>=1.0
pyproj>=3.6
shapely>=2.0
rtree>=1.3
matplotlib>=3.8
contextily>=1.6
streamlit>=1.37
scipy>=1.12
pysal>=24.1
scikit-learn>=1.5
umap-learn>=0.5
xgboost>=2.0
